requirements.txt
Flask
Celery
redis
pytesseract
Pillow  # If needed for image processing
pdf2image  # Add this if you're processing PDF files
